Squat on certain bundle names
To avoid confusion between "official" Operable-authored bundles and 3rd
party bundles, we'll block the installation of bundles named

* `cog`
* `cog-*`
* `operable-*`

The bundle names `site` and `user` continue to be reserved for internal
usage. The embedded bundle continues to be named "operable", and
continues to be uninstallable by outside sources.

We were using "cog" as a bundle name for testing data; obviously that's
out now, so we'll now call it "test-bundle" instead.

Fixes #817